moved config and cleaned up types
