+++ packages/cli/src/tokens/process/platform.ts
@@ -6,8 +6,13 @@ import * as R from 'ramda';
 import StyleDictionary from 'style-dictionary';
 import { cleanDir, readFile, writeFile } from '../../utils.js';
 import type { TokensSet } from '../types.js';
+import {
+  type BuildConfig,
+  type SDConfigForThemePermutation,
+  type ThemePermutation,
+  colorCategories,
+} from '../types.js';
 import { configs, getConfigsForThemeDimensions } from './configs.js';
-import { type BuildConfig, type SDConfigForThemePermutation, type ThemePermutation, colorCategories } from './types.js';
 import { concatFiles } from './utils/entryfile.js';
 import { type ProcessedThemeObject, processThemeObject } from './utils/getMultidimensionalThemes.js';
